refactor(concept-pages): pass through related concepts as simple object properties
Why I did this:  modifying an ember-data model instance just to pass through some extra data from a route model
is not a great practice. Suddenly we are then dealing with model instances that have an extra property which is
not defined in the model.
By simply separating out the arguments to the concept-index-page component we can provide the related concepts
directly, allowing us to merge the relatedFrom and relatedTo relationships per route as desired.